% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootLR.R
\name{print.diagCI}
\alias{print.diagCI}
\title{Prints results from diagCI
As is typical for R, this is run automatically when you type in an object name, and is typically not run directly by the end-user.}
\usage{
\method{print}{diagCI}(x, digits = 3, ...)
}
\arguments{
\item{x}{The diagCI object created by diagCI()}

\item{digits}{Number of digits to round to}

\item{\dots}{Pass-alongs (currently ignored).}
}
\value{
Returns x unaltered.
}
\description{
Prints results from diagCI
As is typical for R, this is run automatically when you type in an object name, and is typically not run directly by the end-user.
}
\examples{
\dontrun{
diagCI( 25, 50, 45, 75 )
}
}
