\name{print.lrtest}
\alias{print.lrtest}
\title{Prints results from the BayesianLR.test}
\usage{
  \method{print}{lrtest} (x, ...)
}
\arguments{
  \item{x}{The lrtest object created by BayesianLR.test}

  \item{\dots}{Pass-alongs (currently ignored)}
}
\value{
  Returns x unaltered
}
\description{
  Prints results from the BayesianLR.test
}

