% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootLR.R
\name{print.lrtest}
\alias{print.lrtest}
\title{Prints results from the BayesianLR.test
As is typical for R, this is run automatically when you type in an object name, and is typically not run directly by the end-user.}
\usage{
\method{print}{lrtest}(x, digits = 3, ...)
}
\arguments{
\item{x}{The lrtest object created by BayesianLR.test.}

\item{digits}{Number of digits to round to for display purposes}

\item{\dots}{Pass-alongs (currently ignored).}
}
\value{
Returns x unaltered.
}
\description{
Prints results from the BayesianLR.test
As is typical for R, this is run automatically when you type in an object name, and is typically not run directly by the end-user.
}
\examples{
\dontrun{
print.lrtest( BayesianLR.test( 500, 500, 300, 500 ), digits = 4 )
}
}

